backend/requirements.txt
fastapi
uvicorn
httpx
SQLAlchemy
python-multipart
httpx
alibabacloud_ocr_api20210707
alibabacloud_tea_openapi
alibabacloud_darabonba_stream
aliyun-python-sdk-core
aliyun-python-sdk-green
python-dotenv